% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_quantiles.R
\name{compute_transmission_quantile}
\alias{compute_transmission_quantile}
\title{Compute Transmission Quantiles}
\usage{
compute_transmission_quantile(q, k, perform_checks = TRUE)
}
\arguments{
\item{q}{A an array of quantiles. Must take values on the unit interval. The proportion of most infectious index cases.}

\item{k}{A positive, real-values scalar. The dispersion parameter for the distribution of secondary infections.}

\item{perform_checks}{Logical. Should function arguments be checked to ensure
they satisfy any constraints.}
}
\value{
An array of values on the unit interval. The proportion  of transmission associated with each elemnt of q.
}
\description{
Compute the expected proportion of transmission associated with the q most infectious index cases
for a given value of the dispersion parameter k. Assumes Negative-Binomial distributed secondary infections.
}
\examples{
compute_transmission_quantile(q = 0.2, k = 0.1)
}
