% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{convert_gamma_moments}
\alias{convert_gamma_moments}
\title{Convert gamma moments}
\usage{
convert_gamma_moments(expectation, standard_deviation, perform_checks = TRUE)
}
\arguments{
\item{expectation}{A scalar. The expected value of the gamma random variable.}

\item{standard_deviation}{A scalar. The standard deviation of the gamma
random variable.}

\item{perform_checks}{Logical. Should function arguments be checked to ensure
they satisfy any constraints.}
}
\value{
A named vector of length 2. The shape and rate for the gamma random
  variable.
}
\description{
Convert the first two moments of a gamma distribution to shape and rate
parameters.
}
\examples{
convert_gamma_moments(5, 1)
}
