% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{convert_inverse_gamma_moments}
\alias{convert_inverse_gamma_moments}
\title{Convert inverse-gamma moments}
\usage{
convert_inverse_gamma_moments(
  expectation,
  standard_deviation,
  perform_checks = TRUE
)
}
\arguments{
\item{expectation}{A scalar. The expected value of the inverse-gamma random variable.}

\item{standard_deviation}{A scalar. The standard deviation of the inverse-gamma
random variable.}

\item{perform_checks}{Logical. Should function arguments be checked to ensure
they satisfy any constraints.}
}
\description{
Convert the first two moments of an inverse-gamma distribution to shape and rate
parameters.
}
\examples{
convert_inverse_gamma_moments(5, 1)
}
