% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homo_hist_Rt_stan.R
\name{homo_hist_Rt_stan}
\alias{homo_hist_Rt_stan}
\title{Bayesian Estimation of time-varying reproduction numbers}
\usage{
homo_hist_Rt_stan(
  epidemic_curve,
  seed_days,
  import_rate,
  generation_interval_mean = 5,
  generation_interval_sd = 2.5,
  generation_interval_length = 21,
  bin_width = 7,
  r_prior_mean = 0,
  r_prior_sd = 1,
  next_day_cases,
  next_day_import_rate,
  c = 1,
  ...
)
}
\arguments{
\item{epidemic_curve}{A vector of integers. The epidemic curve to be
assessed.}

\item{seed_days}{An integer scalar. The number of days included to seed the
epidemic}

\item{import_rate}{A vector of integers. The same length as epidemic_curve.
The rate at which cases are imported from outwith the population.}

\item{generation_interval_mean}{A scalar. The mean of generation intervals
between infector-infectee pairs.}

\item{generation_interval_sd}{A scalar. The standard deviation of  generation
intervals between infector-infectee pairs.}

\item{generation_interval_length}{An integer scalar. The maximum generation
interval.}

\item{bin_width}{An integer scalar. The bin width defining the histogram
estimator for time-varying reproduction numbers.}

\item{r_prior_mean}{A scalar. The mean for a log-normal distribution.}

\item{r_prior_sd}{A positive scalar. The standard deviation for a log-normal
distribution.}

\item{next_day_cases}{An integer scalar. One day ahead case count. Allows
model validation.}

\item{next_day_import_rate}{A positice scalar. Import rate for the day ahead.}

\item{c}{A small positive integer A regularising constant.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian Estimation of time-varying reproduction numbers
}
\examples{
D <- 30
df <- covid_incidence_roi_epidemiological_date[1:D, ]
fit <- homo_hist_Rt_stan(
  epidemic_curve = df$count, seed_days = 5,
  import_rate = rep(1, D),
  next_day_cases =  covid_incidence_roi_epidemiological_date$count[D+1],
  next_day_import_rate = 1
  )
}
