% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_stan.R
\name{initialise_momentum}
\alias{initialise_momentum}
\title{Initialise Momentum.}
\usage{
initialise_momentum(epidemic_curve, k, seed = NULL)
}
\arguments{
\item{epidemic_curve}{A vector of integers. The epidemic curve to be
assessed.}

\item{k}{A positive scalar. The case dispersion parameter. Setting to Inf
allows for homogeneous disease reproduction.}

\item{seed}{A scalar. The random seed.}
}
\value{
A real valued vector. Log disease momentum for each day.
}
\description{
Initialise the momentum in the sampling scheme for heterogeneous disease
reproduction assuming a reproduction number of 1.
}
