% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_stan.R
\name{initialise_lgp_Rt}
\alias{initialise_lgp_Rt}
\title{Initialise sampling scheme.}
\usage{
initialise_lgp_Rt(epidemic_curve, gp_amplitude, k, seed = NULL)
}
\arguments{
\item{epidemic_curve}{A vector of integers. The epidemic curve to be
assessed.}

\item{gp_amplitude}{A positive scalar. The amplitude hyper-parameter for the
underlying Gaussian process.}

\item{k}{A positive scalar. The case dispersion parameter. Setting to Inf
allows for homogeneous disease reproduction.}

\item{seed}{A scalar. The random seed.}
}
\value{
A real valued vector. Log disease momentum for each day.
}
\description{
Initialise the sampling scheme for heterogeneous disease reproduction with a
log-Gaussian process prior on the time-varying reproduction number.
}
