% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_stan.R
\name{initialise_momentum}
\alias{initialise_momentum}
\title{Initialise Momentum.}
\usage{
initialise_momentum(epidemic_curve, log_R_sd = 0.1, k = 0.1, seed = NULL)
}
\arguments{
\item{epidemic_curve}{A vector of integers. The epidemic curve to be
assessed.}

\item{log_R_sd}{A positive scalar. The standard deviation of the zero-mean
log-normal distribution providing initial time-varying reproduction
numbers.}

\item{k}{A positive scalar. The value for the dispersion parameter k from
which initial momentum parameters are sampled.}

\item{seed}{A scalar. The random seed.}
}
\value{
A real valued vector. Log disease momentum for each day.
}
\description{
Initialise the momentum in the sampling scheme for heterogeneous disease
reproduction assuming a reproduction number of 1.
}
